fix(ble/bluedroid): Add null/range checks and fix resource handling in BTA layer
- bta_dm_int: fix BTA_SERVICE_ID_TO_SERVICE_MASK undefined behavior (1<<id when id>=32)
- bta_gattc_main: add event bounds check before state table lookup
- bta_gattc_utils: null checks for remote_bda/p_rcb, fix list_free in clcb_dealloc,
  bta_to_btif_uuid fixes
- bta_gatts_act: fix formatting/indent in send_service_change_indication
- bta_gatts_api: validate attr_val/len, add error logs on alloc failure
- bta_sys_main: null/range checks in sm_execute, alarm/hash_map error handling in
  bta_alarm_cb


(cherry picked from commit f4cec2ac4e2ffa9c74de7ac67d2dacc09945152d)

+++ components/bt/host/bluedroid/bta/gatt/bta_gatts_act.c
@@ -872,7 +872,7 @@ void bta_gatts_send_service_change_indication (tBTA_GATTS_DATA *p_msg)
     tBTA_GATTS_RCB     *p_rcb = bta_gatts_find_app_rcb_by_app_if(p_msg->api_send_service_change.server_if);
     tBTA_GATTS_SERVICE_CHANGE    service_change;
     tBTA_GATT_STATUS status = BTA_GATT_OK;
-    UINT16 addr[BD_ADDR_LEN] = {0};
+    UINT8 addr[BD_ADDR_LEN] = {0};
     if(memcmp(p_msg->api_send_service_change.remote_bda, addr, BD_ADDR_LEN) != 0) {
         BD_ADDR bd_addr;
         memcpy(bd_addr, p_msg->api_send_service_change.remote_bda, BD_ADDR_LEN);
